chore: Updated swagger and changed from int to float64 in implementaion, some failing tests

+++ dto/rockerLaunched.go
@@ -1,7 +1,7 @@
 package DTO
 
 type RocketLaunched struct {
-	LaunchSpeed int    `json:"launchSpeed"`
-	Mission     string `json:"mission"`
-	Type        string `json:"type"`
+	LaunchSpeed float64 `json:"launchSpeed"`
+	Mission     string  `json:"mission"`
+	Type        string  `json:"type"`
 }
